pulsar-functions/ingest-to-astra-function/ingest-to-astra/requirements.txt
python-dotenv
pulsar-client==2.10
astrapy
#dnspython==2.6